backend/requirements.txt
Flask==2.3.3
torch>=2.2.0,<2.3.0
torchvision>=0.17.0,<0.18.0
facenet-pytorch==2.6.0
Pillow>=10.2.0,<10.3.0
matplotlib==3.7.2
pymongo==4.5.0
numpy==1.24.3
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
